Issue #11 - Administrator Logout
This commit contains a feature to test the happy path of logging a user out.
It contains new step definitions to allow for testing of the current page too.
The new specs test that the session is deleted appropriately, and describes
new specs that describe the behaviour of the 'destroy' method of the session
controller when no user is logged in. These are still to be completed.